refactor: Move purge-command
