+++ cmd/monaco/purge/purge.go
@@ -1,18 +1,20 @@
-// @license
-// Copyright 2021 Dynatrace LLC
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+/*
+ * @license
+ * Copyright 2023 Dynatrace LLC
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
-package delete
+package purge
 
 import (
 	"errors"
@@ -21,13 +23,14 @@ import (
 	"github.com/dynatrace/dynatrace-configuration-as-code/internal/log"
 	"github.com/dynatrace/dynatrace-configuration-as-code/internal/maps"
 	"github.com/dynatrace/dynatrace-configuration-as-code/pkg/api"
+	"github.com/dynatrace/dynatrace-configuration-as-code/pkg/client"
 	"github.com/dynatrace/dynatrace-configuration-as-code/pkg/delete"
 	"github.com/dynatrace/dynatrace-configuration-as-code/pkg/manifest"
 	"github.com/spf13/afero"
 	"path/filepath"
 )
 
-func Purge(fs afero.Fs, deploymentManifestPath string, environmentNames []string, apiNames []string) error {
+func purge(fs afero.Fs, deploymentManifestPath string, environmentNames []string, apiNames []string) error {
 
 	deploymentManifestPath = filepath.Clean(deploymentManifestPath)
 	deploymentManifestPath, manifestErr := filepath.Abs(deploymentManifestPath)
@@ -78,7 +81,7 @@ func purgeConfigs(environments []manifest.EnvironmentDefinition, apis map[string
 }
 
 func purgeConfigsForEnvironment(env manifest.EnvironmentDefinition, apis map[string]api.Api) []error {
-	dynatraceClient, err := createClient(env, false)
+	dynatraceClient, err := createClient(env)
 
 	if err != nil {
 		return []error{
@@ -89,4 +92,19 @@ func purgeConfigsForEnvironment(env manifest.EnvironmentDefinition, apis map[str
 	log.Info("Deleting configs for environment `%s`", env.Name)
 
 	return delete.DeleteAllConfigs(dynatraceClient, apis)
+}
+
+func createClient(environment manifest.EnvironmentDefinition) (client.Client, error) {
+	token, err := environment.GetToken()
+
+	if err != nil {
+		return nil, err
+	}
+
+	url, err := environment.GetUrl()
+	if err != nil {
+		return nil, err
+	}
+
+	return client.NewDynatraceClient(url, token)
 }
